requirements.txt
streamlit
supabase
sentence-transformers
anthropic
python-dotenv
watchdog
